requirements.txt
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
